+++ scripts/journal.js
@@ -83,4 +83,5 @@ export async function setupJournalEntry() {
     await preparePages(journalEntry, 'chris-premades.CPR Summon Features');
     await preparePages(journalEntry, 'chris-premades.CPR Actions');
     await preparePages(journalEntry, 'chris-premades.CPR Feat Features');
+    await preparePages(journalEntry, 'chris-premades.CPR Homebrew Feature Items');
 }