+++ scripts/macros/homebrew/bg3/topple.js
@@ -1,0 +1,37 @@
+import {chris} from '../../../helperFunctions.js';
+import {constants} from '../../../constants.js';
+import {queue} from '../../../utility/queue.js';
+async function item({speaker, actor, token, character, item, args, scope, workflow}) {
+    if (workflow.actor.system.abilities.dex.save < workflow.actor.system.abilities.str.save) return;
+    workflow.item = workflow.item.clone({'system.save.scaling': 'dex'}, {'keepId': true});
+}
+async function attack({speaker, actor, token, character, item, args, scope, workflow}) {
+    if (workflow.targets.size != 1 || workflow.item.system.actionType != 'mwak') return;
+    let validTypes = [
+        'quarterstaff'
+    ];
+    if (!validTypes.includes(workflow.item.system.baseItem)) return;
+    let feature = chris.getItem(workflow.actor, 'Topple');
+    if (!feature) return;
+    if (!feature.system.uses.value) return;
+    let queueSetup = await queue.setup(workflow.item.uuid, 'topple', 450);
+    if (!queueSetup) return;
+    let selection = await chris.dialog(feature.name, [['Yes', true], ['No', false]], 'Use Topple?');
+    if (!selection) {
+        queue.remove(workflow.item.uuid);
+        return;
+    }
+    await feature.update({'system.uses.value': 0});
+    let featureData = duplicate(feature.toObject());
+    featureData.system.damage.parts = [];
+    delete (featureData._id);
+    let feature2 = new CONFIG.Item.documentClass(featureData, {'parent': workflow.actor});
+    let [config, options] = constants.syntheticItemWorkflowOptions([workflow.targets.first().document.uuid]);
+    await warpgate.wait(100);
+    await MidiQOL.completeItemUse(feature2, config, options);
+    queue.remove(workflow.item.uuid);
+}
+export let topple = {
+    'item': item,
+    'attack': attack
+}