+++ scripts/macros/monsterFeatures/generic/generic.js
@@ -8,6 +8,7 @@ import {swarmDamage} from './swarm.js';
 import {prone} from './prone.js';
 import {fall} from './fall.js';
 import {underwater} from './underwater.js';
+import {nonLethal} from './nonLethal.js';
 export let generic = {
     'autoGrapple': autoGrapple,
     'swarmDamage': swarmDamage,
@@ -18,5 +19,6 @@ export let generic = {
     'shove': shove,
     'prone': prone,
     'fall': fall,
-    'underwater': underwater
+    'underwater': underwater,
+    'nonLethal': nonLethal
 }