+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # 0.7.19 Release:
+## New Homebrew Content:
+- Baldur's Gate 3 Weapon Actions: Backbreaker, Brace, Cleave, Concussive Smash, Flourish, Hamstring Shot, Heartstopper, Lacerate, Maiming Strike, Mobile Shot, Piercing Shot, Piercing Strike, Pommel Strike, Prepare, Rush Attack, Tenacity, Topple, Weakening Strike
 ## Bug Fixes:
 - Empowered Evocation
 - Tasha's Caustic Brew
 - Sorcery Points
+- Holy Weapon
   
 # 0.7.18 Release:
 ## New Content: